fix type

+++ synthesize.py
@@ -81,7 +81,7 @@ def synthesize(model, input_data, mel_spec, force_cpu=False):
     mean = np.mean(spectrogram, axis=1, keepdims=True)
     std = np.std(spectrogram, axis=1, keepdims=True)
     spectrogram = (spectrogram - mean) / std
-    spectrogram = torch.tensor(np.array([spectrogram]))
+    spectrogram = torch.tensor(np.array([spectrogram]), dtype=torch.float)
     spectrogram = spectrogram.to('cuda')
     s = model.inference(t, speaker=s, target=spectrogram, language=l).cpu().detach().numpy()
     s = audio.denormalize_spectrogram(s, not hp.predict_linear)
